Module limn::prelude

Re-exports of common crate-internal functions / structs

Reexports

pub use cassowary::strength::*;

pub use cassowary::WeightedRelation::*;

pub use geometry::Point;

pub use geometry::Rect;

pub use geometry::RectExt;

pub use geometry::Size;

pub use geometry::SizeExt;

pub use geometry::Vector;

pub use event::EventHandler;

pub use event::EventArgs;

pub use event::event_global;

pub use input::mouse::ClickEvent;

pub use widget::Widget;

pub use widget::draw::Draw;

pub use widget::draw::DrawEventHandler;

pub use widget::property::Property;

pub use widget::property::states::*;

pub use style::Component;

pub use style::DrawState;

pub use style::DrawStyle;

pub use render::RenderBuilder;

pub use resources::resources;

pub use resources::WidgetId;

pub use resources::id::Id;

pub use resources::id::IdGen;

pub use resources::image::ImageSource;

pub use ui::Ui;

pub use app::App;

pub use app::FrameEvent;

pub use window::Window;

pub use color::*;

pub use layout::*;

pub use layout::constraint::*;

pub use maplit::*;

Modules

glutin
The purpose of this library is to provide an OpenGL context on as many platforms as possible.

Macros

btreemap
Create a BTreeMap from a list of key-value pairs

btreeset
Create a BTreeSet from a list of elements.

constraints
Used to specify a list of constraints.

convert_args
Macro that converts the keys or key-value pairs passed to another maplit macro. The default conversion is to use the Into trait, if no custom conversion is passed.

hashmap
Create a HashMap from a list of key-value pairs

hashset
Create a HashSet from a list of elements.

Structs

LAYOUT

LinearLayoutSettings

Enums

Align
The orientation of Alignment along some Axis.

ItemAlignment

Orientation

Spacing

Wrap
The way in which text should wrap around the width.