Trait limn::style::Component

pub trait Component: Clone {
    fn name() -> String;
}

A type that implements Component is any type that can have values stored in and retrieved from the Theme

Required Methods

fn name() -> String