Struct limn::widget::WidgetBuilder

pub struct WidgetBuilder {
    pub widget: WidgetRef,
}

Used to initialize and modify a Widget before it's been added to a parent Widget

Fields
widget: WidgetRef
Methods

impl WidgetBuilder

fn new<S: Into<String>>(name: S) -> Self

Creates a new, named WidgetBuilder, ex. "glcanvas". The WidgetBuilder can then be referred to by name

fn widget_ref(&self) -> WidgetRef

Clones the current widget

fn id(&self) -> WidgetId

Returns the ID of the current widget

fn set_draw_state<T: Draw + 'static>(&mut self, draw_state: T) -> &mut Self

Applies a new Draw-State to the widget.

fn set_draw_state_with_style<T: Draw + 'static, S: Style<T> + 'static>(
&mut self,
draw_state: T,
style: S
) -> &mut Self

Applies both a DrawState and a Style to the current widget

fn add_handler<E: 'static, T: EventHandler<E> + 'static>(
&mut self,
handler: T
) -> &mut Self

Adds a handler to the current widget

fn add_prop(&mut self, property: Property) -> &mut Self

Recursively sets a certain property on the current widget

fn remove_prop(&mut self, property: &Property) -> &mut Self

Recursively removes a certain property on the current widget

fn layout(&mut self) -> LayoutGuardMut

Performs the layout on the current widget

fn add_child<U: Into<WidgetRef>>(&mut self, child: U) -> &mut Self

Adds a child widget to the current Widget. Note that the child may be unconstrained.

fn set_name(&mut self, name: &str) -> &mut Self

Sets the name of the WidgetBuilder

impl WidgetBuilder

fn linear_layout(&mut self, settings: LinearLayoutSettings) -> &mut Self

Set this widgets container to be a LinearLayout. Children added to this widget will be arranged along one axis without overlapping.

fn grid(&mut self, num_columns: usize) -> &mut Self

Set this widgets container to be a GridLayout. Children added to this widget will be arranged in a grid.

impl WidgetBuilder

fn on_click<F>(&mut self, on_click: F) -> &mut Self where
F: Fn(&ClickEvent, &mut EventArgs) + 'static,

impl WidgetBuilder

fn make_draggable(&mut self) -> &mut Self

Make a widget receive drag events.

impl WidgetBuilder

fn list_item(&mut self, parent_list: &WidgetRef) -> &mut Self

fn on_item_selected<F>(&mut self, on_item_selected: F) -> &mut Self where
F: Fn(EventArgs) + 'static,

impl WidgetBuilder

fn enable_hover(&mut self) -> &mut Self

impl WidgetBuilder

fn make_focusable(&mut self) -> &mut Self