Struct limn::widget::Widget

pub struct Widget(_);

Methods

impl Widget

pub fn new<S: Into<String>>(name: S) -> Self

Creates a new, named Widget, ex. "glcanvas". The Widget can then be referred to by name

pub fn from_modifier<T: Component + WidgetModifier>(component: T) -> Self

pub fn from_modifier_style<C: Component + WidgetModifier + 'static, T: ComponentStyle<Component = C> + Debug + Send>(
style: T
) -> Self

pub fn from_modifier_style_class<C: Component + WidgetModifier + 'static, T: ComponentStyle<Component = C> + Debug + Send>(
style: T,
class: &str
) -> Self

pub fn set_draw_state<T: Draw + Component + 'static>(
&mut self,
draw_state: T
) -> &mut Self

pub fn set_draw_style<I: Into<DrawStyle>>(&mut self, style: I) -> &mut Self

pub fn set_cursor_hit_fn<F: Fn(Rect, Point) -> bool + 'static>(
&mut self,
cursor_hit_fn: F
) -> &mut Self

pub fn add_modifier<M: DrawModifier + 'static>(
&mut self,
modifier: M
) -> &mut Self

pub fn add_handler<E: 'static, T: EventHandler<E> + 'static>(
&mut self,
handler: T
) -> &mut Self

pub fn add_handler_fn<E: 'static, T: FnMut(&E, EventArgs) + 'static>(
&mut self,
handler: T
) -> &mut Self

pub fn layout(&mut self) -> LayoutGuardMut

pub fn layout_vars(&self) -> LayoutVars

pub fn props(&self) -> PropsGuard

pub fn add_prop(&mut self, property: Property)

pub fn remove_prop(&mut self, property: Property)

pub fn draw_state(&mut self) -> DrawStateGuard

pub fn downgrade(&self) -> WidgetWeak

pub fn id(&self) -> WidgetId

pub fn set_name(&mut self, name: &str) -> &mut Self

pub fn set_debug_color(&mut self, color: Color) -> &mut Self

pub fn name(&self) -> String

pub fn debug_color(&self) -> Option<Color>

pub fn has_updated(&self) -> bool

pub fn set_updated(&self, has_updated: bool)

pub fn bounds(&self) -> Rect

pub fn update<F, T: Draw + 'static>(&mut self, f: F) where
F: FnOnce(&mut T),

pub fn update_modifier<F, T: DrawModifier + 'static>(&mut self, f: F) where
F: FnOnce(&mut T),

pub fn add_child<U: Into<Widget>>(&mut self, child: U) -> &mut Self

pub fn remove_child(&mut self, child: Widget)

pub fn remove_widget(&mut self)

pub fn parent(&self) -> Option<Widget>

pub fn children(&self) -> Vec<Widget>

pub fn child(&self, name: &str) -> Option<Widget>

pub fn event<T: 'static>(&self, data: T)

pub fn event_subtree<T: 'static>(&self, data: T)

pub fn event_bubble_up<T: 'static>(&self, data: T)

pub fn trigger_event(&self, ui: &mut Ui, type_id: TypeId, event: &Any) -> bool

pub fn draw(&mut self, crop_to: Rect, renderer: &mut RenderBuilder, debug: bool)

pub fn is_under_cursor(&self, cursor: Point) -> bool

impl Widget

pub fn linear_layout(&mut self, settings: LinearLayoutSettings) -> &mut Self

Set this widgets container to be a LinearLayout. Children added to this widget will be arranged along one axis without overlapping.

pub fn grid(&mut self, num_columns: usize) -> &mut Self

Set this widgets container to be a GridLayout. Children added to this widget will be arranged in a grid.

impl Widget

pub fn list_item(&mut self, parent_list: &Widget) -> &mut Self

pub fn on_item_selected<F>(&mut self, on_item_selected: F) -> &mut Self where
F: Fn(EventArgs) + 'static,

pub fn set_contents<C, I, F>(&mut self, contents: C, build: F) where
C: Iterator<Item = I>,
F: Fn(I, &mut Widget) -> Widget,

impl Widget

pub fn enable_hover(&mut self) -> &mut Self

impl Widget

pub fn make_focusable(&mut self) -> &mut Self

impl Widget

pub fn make_draggable(&mut self) -> &mut Self

Make a widget receive drag events.